Ajustes Double e Float

+++ header/Dados.hpp
@@ -3,7 +3,7 @@
 
 typedef unsigned char u1;
 typedef unsigned short u2;
-typedef unsigned int u4;
+typedef unsigned long u4;
 typedef unsigned long long u8;
 
 #define CLASS 1
